Module aml::pci_routing

Re-exports

pub use crate::resource::IrqDescriptor;

Structs

PciRoute

PciRoutingTable

A PciRoutingTable is used to interpret the data in a _PRT object, which provides a mapping from PCI interrupt pins to the inputs of the interrupt controller. One of these objects must be present under each PCI root bridge, and consists of a package of packages, each of which describes the mapping of a single PCI interrupt pin.

Enums

PciRouteType

Pin