Struct aml::pci_routing::PciRoutingTable

pub struct PciRoutingTable { /* private fields */ }

A PciRoutingTable is used to interpret the data in a _PRT object, which provides a mapping from PCI interrupt pins to the inputs of the interrupt controller. One of these objects must be present under each PCI root bridge, and consists of a package of packages, each of which describes the mapping of a single PCI interrupt pin.

Implementations

impl PciRoutingTable

pub fn from_prt_path(
prt_path: &AmlName,
context: &mut AmlContext
) -> Result<PciRoutingTable, AmlError>

Construct a PciRoutingTable from a path to a _PRT object. Returns AmlError::IncompatibleValueConversion if the value passed is not a package, or if any of the values within it are not packages. Returns the various AmlError::Prt* errors if the internal structure of the entries is invalid.

pub fn route(
&self,
device: u16,
function: u16,
pin: Pin,
context: &mut AmlContext
) -> Result<IrqDescriptor, AmlError>

Get the interrupt input that a given PCI interrupt pin is wired to. Returns AmlError::PrtNoEntry if the PRT doesn’t contain an entry for the given address + pin.