Struct aml::AmlName

pub struct AmlName(_);

Implementations

impl AmlName

pub fn root() -> AmlName

pub fn from_name_seg(seg: NameSeg) -> AmlName

pub fn from_components(components: Vec<NameComponent>) -> AmlName

pub fn from_str(string: &str) -> Result<AmlName, AmlError>

Convert a string representation of an AML name into an AmlName.

pub fn as_string(&self) -> String

pub fn is_normal(&self) -> bool

An AML path is normal if it does not contain any prefix elements (“^” characters, when expressed as a string).

pub fn is_absolute(&self) -> bool

pub fn search_rules_apply(&self) -> bool

Special rules apply when searching for certain paths (specifically, those that are made up of a single name segment). Returns true if those rules apply.

pub fn normalize(self) -> Result<AmlName, AmlError>

Normalize an AML path, resolving prefix chars. Returns AmlError::InvalidNormalizedName if the path normalizes to an invalid path (e.g. \^_FOO)

pub fn parent(&self) -> Result<AmlName, AmlError>

Get the parent of this AmlName. For example, the parent of \_SB.PCI0._PRT is \_SB.PCI0. The root path has no parent, and so returns None.

pub fn resolve(&self, scope: &AmlName) -> Result<AmlName, AmlError>

Resolve this path against a given scope, making it absolute. If the path is absolute, it is returned directly. The path is also normalized.