Module aml::value

Structs

Args

A control method can take up to 7 arguments, each of which is an AmlValue.

FieldFlags

MethodFlags

StatusObject

Representation of the return value of a _STA method, which represents the status of an object. It must be evaluated, if present, before evaluating the _INI method for an device.

Enums

AmlType

AmlValue

FieldAccessType

FieldUpdateRule

MethodCode

RegionSpace